Move admin blog handlers to handler package
