+++ server/handler-admin/solution_admin.go
